Making script much better! :)
